requirements.txt
flask
telethon==1.36.0
pyTelegramBotAPI==4.21.0
PyYAML==6.0.1
